// monorepo/apps/web/src/modules/warp/packages/client/features/form/types.ts
export type RecommendationListStateType = {
  questions: QuestionList;
  current: { question: string | null };
};
type QuestionList = {
  QuestionId: string;
  questionKey: string;
}[];
export type RecommendationListActionsType = {
  init: (questions: []) => void;
  nextRecommHandler: (currentQuestionData: any) => string;
  prevRecommHandler: (currentQuestionData: any) => void;
  hasdata: (currentQuestionData: any, isNext: boolean) => [];
};